Remove estimated computation time and clean each time new search happens
Estimated computation is much harder to predict due to the pylon
coverage stuff, so I'm completely removing it (no info > bad info)

Also, clean up after the previous search.
This is mainly noticed when we start a new search, and it takes a long time to
find the first feasible solution. Before this change, the solution set from
the previous search would have remained there until we found the first
feasible solution.
Now, that solution set is immediately cleaned up in the setting up stage
of the new search.